Separated align and score commands
